+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## master
+
+* Add `Video` component. [#176](https://github.com/mapbox/dr-ui/pull/176)
+
 ## 0.20.1
 
 * Add core-js as a dependency.